+++ integration-tests/__tests__/expect-async-matcher.test.js
@@ -10,12 +10,9 @@
 'use strict';
 
 const path = require('path');
-const SkipOnWindows = require('../../scripts/SkipOnWindows');
 const runJest = require('../runJest');
 const {extractSummary} = require('../Utils');
 const dir = path.resolve(__dirname, '../expect-async-matcher');
-
-SkipOnWindows.suite();
 
 test('works with passing tests', () => {
   const result = runJest(dir, ['success.test.js']);