+++ integration-tests/__tests__/test_path_pattern_reporter_message.test.js
@@ -13,10 +13,7 @@ import os from 'os';
 import path from 'path';
 const {cleanup, writeFiles} = require('../Utils');
 
-const SkipOnWindows = require('../../scripts/SkipOnWindows');
 const DIR = path.resolve(os.tmpdir(), 'jest_path_pattern_reporter_message');
-
-SkipOnWindows.suite();
 
 beforeEach(() => cleanup(DIR));
 afterEach(() => cleanup(DIR));